Refactor directory layouts.
1. Move controllers to a dedicated directory, separating it with routes, middlewares, etc.
2. Add response directory to put response structs.
