+++ app/init.go
@@ -1,4 +1,4 @@
-package controller
+package app
 
 import (
 	"github.com/labstack/echo/v4"
