+++ app/response/foo.go
@@ -1,0 +1,6 @@
+package response
+
+type FooResponse struct {
+	Foo string `json:"foo"`
+	Bar int64  `json:"bar"`
+}
