+++ app/routes.go
@@ -1,7 +1,8 @@
-package controller
+package app
 
 import (
 	echoSwagger "github.com/swaggo/echo-swagger"
+	"go-svc-tpl/app/controller"
 )
 
 func addRoutes() {
@@ -9,5 +10,5 @@ func addRoutes() {
 
 	api.GET("/doc/*", echoSwagger.WrapHandler)
 
-	api.GET("/foo", foo)
+	api.GET("/foo", controller.Foo)
 }
